+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to the "php-docblocker" extension will be documented in this
 - Supported fully qualifies class namespace use with bracket `use some\namespace\{ ClassA, ClassB, ... }`
 - Supported fully qualifies class namespace use with comma `use some\namespace\ClassA, some\namespace\ClassB, ...`
 - Fix issue with coverage reports not firing
+- Supported PHP 8.1 Readonly Properties
+- Supported PHP 8.1 Intersection Types
 
 ## [2.6.1] - 2021-10-12
 - Fix double start delimeter when vscode setting `editor.autoClosingBrackets` is set to `never`